break up manage-account.js into more sensible chunks
